[#154784173] moved PageHeaderSimple to Organisms
