+++ spec/javascript/components/organisms/PageHeaderSimple.test.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
-import PageHeaderSimple from 'components/molecules/PageHeaderSimple'
+import PageHeaderSimple from 'components/organisms/PageHeaderSimple'
 import sinon from 'sinon'
 
 describe('PageHeaderSimple', () => {
   let title = "title1"
-  
+
   test('should render succesfully', () => {
     const component = renderer.create(
       <PageHeaderSimple title={title} />,
